Tests are now PSR2
