+++ tests/Fixture/TempHandler.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace rcrowe\Raven\Tests\Fixture;
+
+use rcrowe\Raven\Handler\BaseHandler;
+
+class TempHandler extends BaseHandler
+{
+    public function process($url, $data, array $headers = array())
+    {
+        return true;
+    }
+}
